Individual template refactor
